pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.0",
    "pycryptodome==3.19.0",
    "pyotp==2.9.0",
    "requests==2.31.0",
]
